+++ test/sources/Misc.sol
@@ -1,0 +1,9 @@
+pragma solidity ^0.5.1;
+
+contract Misc {
+
+    string misc;
+
+    constructor() public {
+    }
+}