+++ CHANGELOG.md
@@ -90,3 +90,4 @@ Released with 1.0.0-beta.37 code base.
 ### Fixed
 
 - Fix TS types for eth.subscribe syncing, newBlockHeaders, pendingTransactions (#3159)
+- Improve web3-eth-abi decodeParameters error message (#3134)